Add club profile management views and functionality
- Created edit and index views for club profiles, allowing clubs to update their information and view details.
- Implemented scoring index view for live match scoring and fixture management.
- Added squad management views including create, edit, and index for squads.
- Introduced landing page for the application with features overview.
- Updated routes for clubs, players, and public match APIs to support new functionalities.
- Enhanced squad form with color selection and active status toggle.
- Implemented confirmation dialog for squad deletion.

+++ app/Http/Controllers/Api/ProfileController.php
@@ -433,7 +433,7 @@ class ProfileController extends Controller
             'invited_user_id' => $invitee?->id,
             'role' => $validated['role'] ?? 'player',
             'status' => 'pending',
-            'expires_at' => now()->addDays($validated['expires_in_days'] ?? 7),
+            'expires_at' => now()->addDays((int) ($validated['expires_in_days'] ?? 7)),
             'message' => $validated['message'] ?? null,
         ]);
